third_party/blink/web_tests/platform/generic/fast/events/touch/touch-slider-no-js-touch-listener-expected.txt

Tests to ensure that touch events are delivered to an input element with type=range even when there are no touch event handlers in Javascript. This test is only expected to pass if ENABLE_TOUCH_SLIDER is defined.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS slider.value is "50"
PASS slider.value is "0"
PASS successfullyParsed is true

TEST COMPLETE

PASS successfullyParsed is true

TEST COMPLETE
